Move EndpointIndex logic to common area, use in ConfigGenTest
This makes no change in prod code, and almost no change in test code. It
is a dependency for https://github.com/istio/istio/pull/46329, though.
Once that lands, we need a proper XDSUpdater keeping the EndpointIndex
in sync for all cases; today the tets under v1alpha3/ do not use a
DiscoveryServer (can cannot - looping deps), but they need an
EndpointIndex. This prepares them to use one.
