+++ pilot/pkg/xds/eds_test.go
@@ -648,91 +648,6 @@ func TestDeleteService(t *testing.T) {
 	}
 }
 
-var (
-	c1Key = model.ShardKey{Cluster: "c1"}
-	c2Key = model.ShardKey{Cluster: "c2"}
-)
-
-func TestUpdateServiceAccount(t *testing.T) {
-	cluster1Endppoints := []*model.IstioEndpoint{
-		{Address: "10.172.0.1", ServiceAccount: "sa1"},
-		{Address: "10.172.0.2", ServiceAccount: "sa-vm1"},
-	}
-
-	testCases := []struct {
-		name      string
-		shardKey  model.ShardKey
-		endpoints []*model.IstioEndpoint
-		expect    bool
-	}{
-		{
-			name:      "added new endpoint",
-			shardKey:  c1Key,
-			endpoints: append(cluster1Endppoints, &model.IstioEndpoint{Address: "10.172.0.3", ServiceAccount: "sa1"}),
-			expect:    false,
-		},
-		{
-			name:      "added new sa",
-			shardKey:  c1Key,
-			endpoints: append(cluster1Endppoints, &model.IstioEndpoint{Address: "10.172.0.3", ServiceAccount: "sa2"}),
-			expect:    true,
-		},
-		{
-			name:     "updated endpoints address",
-			shardKey: c1Key,
-			endpoints: []*model.IstioEndpoint{
-				{Address: "10.172.0.5", ServiceAccount: "sa1"},
-				{Address: "10.172.0.2", ServiceAccount: "sa-vm1"},
-			},
-			expect: false,
-		},
-		{
-			name:     "deleted one endpoint with unique sa",
-			shardKey: c1Key,
-			endpoints: []*model.IstioEndpoint{
-				{Address: "10.172.0.1", ServiceAccount: "sa1"},
-			},
-			expect: true,
-		},
-		{
-			name:     "deleted one endpoint with duplicate sa",
-			shardKey: c1Key,
-			endpoints: []*model.IstioEndpoint{
-				{Address: "10.172.0.2", ServiceAccount: "sa-vm1"},
-			},
-			expect: false,
-		},
-		{
-			name:      "deleted endpoints",
-			shardKey:  c1Key,
-			endpoints: nil,
-			expect:    true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			s := new(xds.DiscoveryServer)
-			originalEndpointsShard := &model.EndpointShards{
-				Shards: map[model.ShardKey][]*model.IstioEndpoint{
-					c1Key: cluster1Endppoints,
-					c2Key: {{Address: "10.244.0.1", ServiceAccount: "sa1"}, {Address: "10.244.0.2", ServiceAccount: "sa-vm2"}},
-				},
-				ServiceAccounts: map[string]struct{}{
-					"sa1":    {},
-					"sa-vm1": {},
-					"sa-vm2": {},
-				},
-			}
-			originalEndpointsShard.Shards[tc.shardKey] = tc.endpoints
-			ret := s.UpdateServiceAccount(originalEndpointsShard, "test-svc")
-			if ret != tc.expect {
-				t.Errorf("expect UpdateServiceAccount %v, but got %v", tc.expect, ret)
-			}
-		})
-	}
-}
-
 func fullPush(s *xds.FakeDiscoveryServer) {
 	s.Discovery.Push(&model.PushRequest{Full: true})
 }
